IsShortArrayOrList: add performance tests
These tests do not test any specific syntax combination for the "is short array/short list" determination, they test the performance in combination with a huge array with the some of the most ambivalent list vs array syntax possible, instead.
Additionally, these tests will be testing the impact of the caching which will be implemented.

These tests have all been annotated with `@small`, meaning: they are all expected to return a result in less than 1 second, though in reality, once the improvements have been put in place, they should be much faster than that still.

With the `PHP_Invoker` package + the pcntl PHP extension + the `--enforce-time-limit` CLI option + the `failOnRisky="true"` setting in place as per the previous commit, this should fail the tests in CI when the performance of the `IsShortArrayOrList` class would degrade.